requirements.txt
Django==5.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
django-cors-headers==4.7.0
django-filter==25.1
drf-yasg==1.21.10
python-decouple==3.8
psycopg2-binary==2.9.10
gunicorn==23.0.0
Pillow==11.2.1